mongo/pyproject.toml
[project]
name = "mongo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymongo>=4.6",
    "python-dotenv>=1.0",
    "websockets>=12",
]
